packages/montty/montty-1.0.3.tar.gz/montty-1.0.3/src/montty/io/dir/bash_scripts_dir.py
# MIT License
#
# Copyright (c) 2024-2025 Gwyn Davies
#
# Permission is hereby granted, free of charge, to any person obtaining a copy
# of this software and associated documentation files (the "Software"), to deal
# in the Software without restriction, including without limitation the rights
# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
# copies of the Software, and to permit persons to whom the Software is
# furnished to do so, subject to the following conditions:
#
# The above copyright notice and this permission notice shall be included in all
# copies or substantial portions of the Software.
#
# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR,
# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
# SOFTWARE.

import os
from montty.io.dir.checks_dir import ChecksDir
from montty.platform.config_values import ConfigValues


class BashScriptsDir(ChecksDir):
    def __init__(self):
        super().__init__()
        self._path = f"{super().get_path()}/{ConfigValues.CHECKS_DIR_BASH_SCRIPTS_DIR_NAME}"

    def get_path(self):
        return self._path

    def exists(self):
        return os.path.exists(self.get_path())

    def get_full_path(self, file_name):
        return f"{self.get_path()}/{file_name}"
